create point-adapter and restructure api services
