+++ src/adapters/point-adapter.js
@@ -1,0 +1,35 @@
+export default class PointAdapter {
+  static adaptToServer(point) {
+    const adaptedPoint = {
+      ...point,
+      'base_price' : point.basePrice,
+      'date_from'  : point.dateFrom instanceof Date ? point.dateFrom.toISOString() : null,
+      'date_to'    : point.dateTo instanceof Date ? point.dateTo.toISOString() : null,
+      'is_favorite': point.isFavorite,
+    };
+
+    delete adaptedPoint.basePrice;
+    delete adaptedPoint.dateFrom;
+    delete adaptedPoint.dateTo;
+    delete adaptedPoint.isFavorite;
+
+    return adaptedPoint;
+  }
+
+  static adaptToClient(point) {
+    const adaptedPoint = {
+      ...point,
+      basePrice : point['base_price'],
+      dateFrom  : point['date_from'] !== null ? new Date(point['date_from']) : point['date_from'],
+      dateTo    : point['date_to'] !== null ? new Date(point['date_to']) : point['date_to'],
+      isFavorite: point['is_favorite'],
+    };
+
+    delete adaptedPoint['base_price'];
+    delete adaptedPoint['date_from'];
+    delete adaptedPoint['date_to'];
+    delete adaptedPoint['is_favorite'];
+
+    return adaptedPoint;
+  }
+}
