+++ src/models/points-model.js
@@ -1,3 +1,4 @@
+import PointAdapter from '../adapters/point-adapter.js';
 import { UpdateType } from '../const.js';
 import Observable from '../framework/observable.js';
 
@@ -17,7 +18,7 @@ export default class PointsModel extends Observable {
   async init() {
     try {
       const points = await this.#pointApiService.points;
-      this.#points = points.map(this.#adaptToClient);
+      this.#points = points.map(PointAdapter.adaptToClient);
     } catch(err) {
       this.#points = [];
     }
@@ -34,7 +35,7 @@ export default class PointsModel extends Observable {
 
     try {
       const response = await this.#pointApiService.updatePoint(update);
-      const updatedPoint = this.#adaptToClient(response);
+      const updatedPoint = PointAdapter.adaptToClient(response);
       this.#points = [
         ...this.#points.slice(0, index),
         updatedPoint,
@@ -49,7 +50,7 @@ export default class PointsModel extends Observable {
   async addPoint(updateType, update) {
     try {
       const response = await this.#pointApiService.addPoint(update);
-      const newPoint = this.#adaptToClient(response);
+      const newPoint = PointAdapter.adaptToClient(response);
       this.#points = [
         newPoint,
         ...this.#points,
@@ -76,21 +77,4 @@ export default class PointsModel extends Observable {
     }
   }
 
-  // TODO вынести этот метод в отдельный класс адаптер
-  #adaptToClient(point) {
-    const adaptedPoint = {
-      ...point,
-      basePrice : point['base_price'],
-      dateFrom  : point['date_from'] !== null ? new Date(point['date_from']) : point['date_from'],
-      dateTo    : point['date_to'] !== null ? new Date(point['date_to']) : point['date_to'],
-      isFavorite: point['is_favorite'],
-    };
-
-    delete adaptedPoint['base_price'];
-    delete adaptedPoint['date_from'];
-    delete adaptedPoint['date_to'];
-    delete adaptedPoint['is_favorite'];
-
-    return adaptedPoint;
-  }
 }
